Non-matching decomp of func_800BC760_CB710

+++ include/functions.us.h
@@ -745,7 +745,7 @@ void func_800DC18C_EB13C(Vec3f *arg0, u8 *arg1, u8 *arg2, u16 arg3, u8 arg4);
 void func_800DC5B8_EB568(Vec3f *arg0, f32 arg1, s32 arg2, s32 arg3);
 void func_800B2354_C1304(s16 *arg0, s8 *arg1, s32 arg2, s16 arg3);
 void func_800D702C_E5FDC(s16 arg0, s32 arg1);
-void func_800BC760_CB710(s16 arg0, s16 arg1, s16 arg2, u8 arg3, s32 arg4);
+void func_800BC760_CB710(s16 arg0, s16 arg1, s16 arg2, u8 arg3, s16 arg4);
 void func_800E94B8_F8468(void);
 s32 func_800E95BC_F856C(s32 arg0, s32 arg1, s32 arg2);
 Gfx* func_800E88C0_F7870(s32 arg0, s32 arg1);
